🔥 ✨ Remove react router native and add react navigation, remove a lot codes
