+++ app.json
@@ -1,7 +1,7 @@
 {
   "expo": {
-    "name": "AutismProject",
-    "slug": "AutismProject",
+    "name": "projectbruno",
+    "slug": "projectbruno",
     "version": "1.0.0",
     "orientation": "portrait",
     "icon": "./assets/icon.png",
@@ -27,7 +27,6 @@
     },
     "web": {
       "favicon": "./assets/favicon.png"
-    },
-    "description": ""
+    }
   }
 }
