+++ package.json
@@ -9,13 +9,15 @@
   },
   "dependencies": {
     "@react-native-community/audio-toolkit": "^2.0.3",
+    "@react-navigation/native": "^5.9.3",
+    "@react-navigation/stack": "^5.14.3",
     "expo": "~40.0.0",
     "expo-av": "~8.7.0",
     "expo-status-bar": "~1.0.3",
     "react": "16.13.1",
     "react-dom": "16.13.1",
     "react-native": "https://github.com/expo/react-native/archive/sdk-40.0.1.tar.gz",
-    "react-native-parallax": "^0.3.0",
+    "react-native-screens": "^2.18.1",
     "react-native-web": "~0.13.12",
     "styled-components": "^5.2.1"
   },
