+++ pages/animals/assets/images/cat.png
