+++ pages/animals/styles.js
