+++ pages/home/styles.js
@@ -1,0 +1,8 @@
+import styled from "styled-components/native";
+
+export const Container = styled.View`
+    display: flex;
+`;
+export const TouchableOpacity = styled.View`
+    background-color: blue, height 50%, width 50%;
+`;
