+++ pages/padrin/index.js
@@ -1,0 +1,12 @@
+import React from "react";
+import { Container } from "./styles";
+
+function padrin() {
+    return (
+        <Container>
+            <h1>the best padrin of the world</h1>
+        </Container>
+    );
+}
+
+export default padrin;
